Add post-parsing verify function
Allow user to specify a function to run on the parsed output.  This
function gets the req and res as does the normal parsing middlewares but
operates on the parsed data.

This patch is only a sketch to illustrate my idea, I haven't implemented
tests or anything, but wanted to see if such a feature is of interest to
the project.

My specific usecase is that i want to use bodyparser with json-schema.
I wanted to use the verify function option, but in order to do that I
would need to reparse the json.  I could also wrap the bodyparser
middleware in my own middleware, but this is more complicated than
adding for example "verifyParsed: (rq,rs,b) => {validate(b)}" to the
bodyparser middleware.